move multi-view enum and delegates into subdir; update imports
