+++ src/viewProviders/flinkArtifacts.ts
@@ -1,20 +1,21 @@
-import { Disposable, TreeItem } from "vscode";
+import { Disposable } from "vscode";
 import { ContextValues } from "../context/values";
 import { currentFlinkArtifactsPoolChanged, flinkArtifactUDFViewMode } from "../emitters";
-import { isResponseError, logError } from "../errors";
-import { CCloudResourceLoader } from "../loaders";
-import { FlinkArtifact, FlinkArtifactTreeItem } from "../models/flinkArtifact";
+import { logError } from "../errors";
+import { FlinkArtifact } from "../models/flinkArtifact";
 import { CCloudFlinkComputePool } from "../models/flinkComputePool";
-import { FlinkUdf, FlinkUdfTreeItem } from "../models/flinkUDF";
+import { FlinkUdf } from "../models/flinkUDF";
 import { showErrorNotificationWithButtons } from "../notifications";
 import { MultiModeViewProvider, ViewProviderDelegate } from "./baseModels/multiViewBase";
+import { FlinkArtifactsViewProviderMode } from "./multiViewDelegates/constants";
+import { FlinkArtifactsDelegate } from "./multiViewDelegates/flinkArtifactsDelegate";
+import { FlinkUDFsDelegate } from "./multiViewDelegates/flinkUDFsDelegate";
 
-export enum FlinkArtifactsViewProviderMode {
-  Artifacts = "artifacts",
-  UDFs = "UDFs",
-}
-
-/** Multi-mode view provider for Flink artifacts and UDFs. */
+/**
+ * Multi-mode view provider for Flink artifacts and UDFs.
+ * - When set to the "artifacts" mode, logic is delegated to the {@link FlinkArtifactsDelegate}.
+ * - When set to the "udfs" mode, logic is delegated to the {@link FlinkUDFsDelegate}.
+ */
 export class FlinkArtifactsUDFsViewProvider extends MultiModeViewProvider<
   FlinkArtifactsViewProviderMode,
   CCloudFlinkComputePool,
@@ -33,6 +34,7 @@ export class FlinkArtifactsUDFsViewProvider extends MultiModeViewProvider<
     // the provider itself and causing circular dependencies / stack overflows
     const artifactsDelegate = new FlinkArtifactsDelegate(this);
     const udfsDelegate = new FlinkUDFsDelegate(this);
+
     this.treeViewDelegates = new Map<
       FlinkArtifactsViewProviderMode,
       ViewProviderDelegate<FlinkArtifactsViewProviderMode, FlinkArtifact | FlinkUdf>
@@ -40,6 +42,7 @@ export class FlinkArtifactsUDFsViewProvider extends MultiModeViewProvider<
       [FlinkArtifactsViewProviderMode.Artifacts, artifactsDelegate],
       [FlinkArtifactsViewProviderMode.UDFs, udfsDelegate],
     ]);
+
     this.defaultDelegate = artifactsDelegate;
     this.currentDelegate = this.defaultDelegate;
   }
@@ -85,134 +88,4 @@ export class FlinkArtifactsUDFsViewProvider extends MultiModeViewProvider<
   get computePool(): CCloudFlinkComputePool | null {
     return this.resource;
   }
-}
-
-export class FlinkArtifactsDelegate extends ViewProviderDelegate<
-  FlinkArtifactsViewProviderMode,
-  FlinkArtifact
-> {
-  readonly mode = FlinkArtifactsViewProviderMode.Artifacts;
-  readonly viewTitle = "Flink Artifacts (Preview)";
-
-  children: FlinkArtifact[] = [];
-
-  loadingMessage = "Loading Flink artifacts...";
-
-  constructor(readonly parent: FlinkArtifactsUDFsViewProvider) {
-    super();
-  }
-
-  async fetchChildren(): Promise<FlinkArtifact[]> {
-    this.children = [];
-    try {
-      const loader = CCloudResourceLoader.getInstance();
-      this.children = await loader.getFlinkArtifacts(this.parent.computePool!);
-      return this.children;
-    } catch (error) {
-      const { showNotification, message } = triageGetFlinkArtifactsError(error, this.parent.logger);
-      if (showNotification) {
-        void showErrorNotificationWithButtons(message);
-      }
-      throw error;
-    }
-  }
-
-  getTreeItem(element: FlinkArtifact): TreeItem {
-    return new FlinkArtifactTreeItem(element);
-  }
-}
-
-export class FlinkUDFsDelegate extends ViewProviderDelegate<
-  FlinkArtifactsViewProviderMode,
-  FlinkUdf
-> {
-  readonly mode = FlinkArtifactsViewProviderMode.UDFs;
-  readonly viewTitle = "Flink UDFs (Preview)";
-
-  children: FlinkUdf[] = [];
-
-  loadingMessage = "Loading Flink UDFs...";
-
-  constructor(readonly parent: FlinkArtifactsUDFsViewProvider) {
-    super();
-  }
-
-  async fetchChildren(): Promise<FlinkUdf[]> {
-    this.children = [];
-
-    if (this.parent.computePool) {
-      // TODO: replace this when https://github.com/confluentinc/vscode/issues/2310 is done
-      this.children = [
-        new FlinkUdf({
-          connectionId: this.parent.computePool!.connectionId,
-          connectionType: this.parent.computePool!.connectionType,
-          environmentId: this.parent.computePool!.environmentId,
-          id: "example-udf",
-          name: "Example UDF",
-          description: "This is an example UDF for demonstration purposes.",
-          provider: this.parent.computePool!.provider,
-          region: this.parent.computePool!.region,
-        }),
-      ];
-    }
-    return this.children;
-  }
-
-  getTreeItem(element: FlinkUdf): TreeItem {
-    return new FlinkUdfTreeItem(element);
-  }
-}
-
-export function triageGetFlinkArtifactsError(
-  error: unknown,
-  logger: { debug: (msg: string, err: unknown) => void },
-): {
-  showNotification: boolean;
-  message: string;
-} {
-  let showNotification = false;
-  let message = "Failed to load Flink artifacts.";
-
-  if (isResponseError(error)) {
-    const status = error.response.status;
-    error.response
-      .clone()
-      .json()
-      .catch((err) => {
-        logger.debug("Failed to parse error response as JSON", err);
-      });
-    /* Note: This switch statement intentionally excludes 400 errors.
-     Otherwise, they may pop up on loading the compute pool if it is using an unsupported cloud provider. */
-    if (status >= 401 && status < 600) {
-      showNotification = true;
-      switch (status) {
-        case 401:
-          message = "Authentication required to load Flink artifacts.";
-          break;
-        case 403:
-          message = "Failed to load Flink artifacts. Please check your permissions and try again.";
-          break;
-        case 404:
-          message = "Flink artifacts not found for this compute pool.";
-          break;
-        case 429:
-          message = "Too many requests. Please try again later.";
-          break;
-        case 503:
-          message =
-            "Failed to load Flink artifacts. The service is temporarily unavailable. Please try again later.";
-          break;
-        default:
-          message = "Failed to load Flink artifacts due to an unexpected error.";
-          break;
-      }
-    }
-    void logError(error, "Failed to load Flink artifacts");
-  } else {
-    message = "Failed to load Flink artifacts. Please check your connection and try again.";
-    showNotification = true;
-    void logError(error, "Failed to load Flink artifacts");
-  }
-
-  return { showNotification, message };
 }
