+++ src/viewProviders/multiViewDelegates/constants.ts
@@ -1,0 +1,4 @@
+export enum FlinkArtifactsViewProviderMode {
+  Artifacts = "artifacts",
+  UDFs = "UDFs",
+}
